compress analytics data when storing to dynamo
Uses GZIP compression on the son string and writes as a binary field. This means that the to / from json approach for creating objects doesn't work any more so these are constructed the old fashioned way.

On read the compressed data field is checked, if it exists then use this if not assume the entry is just storing a JSON string and use that.