various ga page view report changes
adds sampling warning when sampling was used to generate results. This appears in the logs and is not shown to real users

uses the labs account for date based queries after 12 dec 16 (when it was created) as this doesn't sample since it contains less data (but still all the data we care about)

paid campaigns now look up by section not individual content - this is sampled before 12/12 but the sample rate in the v4 api is about 10% and results look good - this has been verified with Rob Freeman. By doing this we pull in section fronts and and campaign tracked content for paid content campaigns, which is technically correct.

use page title not url to group data for content, this mean campaign tracking in urls does not cause new entries to be created. NB this causes sampling prior to 12/12 but results still look sensible (graphs are the same shape etc), there is about a 9% variance caused by the sampling, again this effects historic data only and has been confirmed not to be a problem.