Add benchmarks to Campaign Central.
Benchmarks is a new page, requested by Chloe, that looks to provide the metrics we provide at campaign level across ALL campaigns. This at current is done manually in a spreadsheet and repeated by one or more people. This information is generally used to give an idea of what sales can use to sell and is informative in that process.